% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{add_extension}
\alias{add_extension}
\title{Adds file extension to file names}
\usage{
add_extension(path, ext)
}
\arguments{
\item{path}{string or vector of strings of file names}

\item{ext}{string new extension, with or without the dot}
}
\value{
string or vector of strings of file names without extensions
}
\description{
Adds file extension to file names
}
