% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{basename}
\alias{basename}
\title{Overloads basename with option to remove extensions}
\usage{
basename(path, remove.extension = FALSE)
}
\arguments{
\item{path}{string or vector of strings of file names}

\item{remove.extension}{logical whether to remove extension (including dot),
defaults to FALSE}
}
\description{
Overloads basename with option to remove extensions
}
\examples{
basename("/dir1/dir2/file.ext", remove.extension = TRUE)
basename("/dir1/dir2/file.name.ext", remove.extension = TRUE)
basename("/dir1/dir2/file.ext")

}
