% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{change_extension}
\alias{change_extension}
\title{Change file extension, or add if it is missing}
\usage{
change_extension(path, ext)
}
\arguments{
\item{path}{string or vector of strings of file names}

\item{ext}{string new extension, with or without the dot}
}
\value{
string or vector of strings of modified file names
}
\description{
Change file extension, or add if it is missing
}
