% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats.R
\name{count_by}
\alias{count_by}
\title{Count number of observations in a data.table summarized by a certain set of fields}
\usage{
count_by(dt, fields)
}
\arguments{
\item{dt}{data.table to summarize}

\item{fields}{string or vector of strings, sets of fields to summarize the observations by.
Use "-" between column names to summarize by multiple fields. Use a single "*" to return
the unconditional number of rows of the data.table}
}
\description{
Count number of observations in a data.table summarized by a certain set of fields
}
\examples{
iris_dt <- as.data.table(iris)
count_by(iris_dt, c("*", "Species", "Species-Petal.Width"))

}
