% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2.R
\name{create_spot_instances}
\alias{create_spot_instances}
\title{Shortcut function for launching EC2 spot instances}
\usage{
create_spot_instances(instance_params, conf = list(), spot_options = list())
}
\arguments{
\item{instance_params}{list with instance_type, ami, n_instances, pem_file, security_group, subnet}

\item{conf}{list with 'key', 'secret', and 'region' parameters. If missing, will
look for AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY and AWS_DEFAULT_REGION
in environment variables}

\item{spot_options}{list with optional spot instance options}
}
\description{
Shortcut function for launching EC2 spot instances
}
