% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{dirname}
\alias{dirname}
\title{Overloads dirname with option to remove URL scheme ('http://', etc.)
Also changes dirname behavior so that the last part of a URL ending in /
counts as a directory, not as a file}
\usage{
dirname(path, remove.scheme = FALSE)
}
\arguments{
\item{path}{string or vector of strings of file names}

\item{remove.scheme}{logical whether to remove URL scheme,
defaults to FALSE}
}
\description{
Overloads dirname with option to remove URL scheme ('http://', etc.)
Also changes dirname behavior so that the last part of a URL ending in /
counts as a directory, not as a file
}
\examples{
dirname("/dir1/dir2/file.ext", remove.scheme = TRUE)
dirname("/dir1/dir2/file.ext")

}
