% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{extension}
\alias{extension}
\title{Get file extension from file name}
\usage{
extension(path, dot = TRUE)
}
\arguments{
\item{path}{string or vector of strings of file names}

\item{dot}{logical whether to include dot with extension (defaults to TRUE)
No dot if extension is missing}
}
\value{
string or vector of string of file extensions
}
\description{
Get file extension from file name
}
