% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{extract_ymd}
\alias{extract_ymd}
\title{Find and extract date from string with regular expression}
\usage{
extract_ymd(x)
}
\arguments{
\item{x}{string source string to extract date from}
}
\description{
Find and extract date from string with regular expression
}
\examples{
extract_ymd("some_string_2020-04-03-nokey")
extract_ymd("another20100505-02-04")
extract_ymd("nodatehere")

}
