% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table.R
\name{factor_column}
\alias{factor_column}
\title{Turn column to factor by reference}
\usage{
factor_column(dt, column_name)
}
\arguments{
\item{dt}{data.table source string to extract date from}
}
\description{
Preserves attribute "label" of the column if it exists
}
\examples{
dt <- data.table(id=c("A", "A", "B"), value=c(1, 2, 3))
factor_column(dt, "id")

}
