% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_hostname}
\alias{get_hostname}
\title{Get hostname from URL}
\usage{
get_hostname(url)
}
\arguments{
\item{url}{string or vector of strings of URLs}
}
\description{
Get hostname from URL
}
\examples{
get_path("http://example.com")
get_path("https://www.example.org/index.html")
get_path("s3://hfty-test-bucket/dir1/file.txt")

}
