% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2.R
\name{get_instances_ip}
\alias{get_instances_ip}
\title{Collect IP addresses from EC2 instances}
\usage{
get_instances_ip(ec2_instances, conf = list())
}
\arguments{
\item{ec2_instances}{list of EC2 instances returned by aws.ec2}

\item{conf}{list with 'key', 'secret', and 'region' parameters. If missing, will
look for AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY and AWS_DEFAULT_REGION
in environment variables}
}
\description{
Collect IP addresses from EC2 instances
}
