% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_path}
\alias{get_path}
\title{Get path from URL}
\usage{
get_path(url)
}
\arguments{
\item{url}{string or vector of strings of URLs}
}
\description{
Get path from URL
}
\examples{
get_path("http://example.com")
get_path("https://www.example.org/index.html")
get_path("s3://hfty-test-bucket/dir1/file.txt")

}
