% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_url_scheme}
\alias{get_url_scheme}
\title{Get URL scheme ('http://', etc.)}
\usage{
get_url_scheme(url, colonslashes = TRUE)
}
\arguments{
\item{url}{string or vector of strings of URLs}

\item{colonslashes}{whether to include '://' and ':', defaults to TRUE
Not included when scheme is missing}
}
\description{
Get URL scheme ('http://', etc.)
}
\examples{
get_url_scheme("http://example.com")
get_url_scheme("https://www.example.org/index.html")
get_url_scheme("s3://hfty-test-bucket/dir1/file.txt", colonslashes = FALSE)
}
