% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{list_files}
\alias{list_files}
\title{Wraps list.files with better defaults (all files, full paths,
recursive, no dots)}
\usage{
list_files(
  path = ".",
  pattern = NULL,
  all.files = TRUE,
  full.names = TRUE,
  recursive = TRUE,
  ignore.case = FALSE,
  include.dirs = FALSE,
  no.. = TRUE
)
}
\arguments{
\item{path}{vector of full path names; defaults to working directory}

\item{pattern}{optional regular expression to filter paths}

\item{all.files}{logical, default TRUE}

\item{full.names}{logical, default TRUE}

\item{recursive}{logical, default TRUE}

\item{ignore.case}{logical, default FALSE}

\item{include.dirs}{logical, default FALSE}

\item{no..}{logical, default TRUE}
}
\description{
Wraps list.files with better defaults (all files, full paths,
recursive, no dots)
}
\examples{
list_files(".")

}
