% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec2.R
\name{list_running_instances}
\alias{list_running_instances}
\title{List running instances}
\usage{
list_running_instances(conf = list())
}
\arguments{
\item{conf}{list with 'key', 'secret', and 'region' parameters. If missing, will
look for AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY and AWS_DEFAULT_REGION
in environment variables}
}
\description{
List running instances
}
