% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{make_list_of_size}
\alias{make_list_of_size}
\title{Pad list to a certain size with NAs or another value}
\usage{
make_list_of_size(x, size, pad = NA_real_)
}
\arguments{
\item{x}{vector or list}

\item{size}{target size after padding}

\item{pad}{what to pad with, defaults to NA_real_}
}
\description{
Pad list to a certain size with NAs or another value
}
\examples{
make_list_of_size(c(1, 2), 4)

}
