% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{msg}
\alias{msg}
\title{Extends message with quiet option
You may use multiple objects like in message(), but they need to be passed after quiet = FALSE}
\usage{
msg(object = NULL, quiet = FALSE, ...)
}
\arguments{
\item{object}{object passed to message(), default to NULL}

\item{quiet}{logical turn off output messages, default to FALSE}

\item{...}{additional objects or parameters passed to message()}
}
\description{
Extends message with quiet option
You may use multiple objects like in message(), but they need to be passed after quiet = FALSE
}
\examples{
msg("Usage note")
msg(quiet = FALSE, "Usage note: ", "please wash your hands")
msg(quiet = TRUE, "Usage note: ", "please wash your hands")

}
