% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectors.R
\name{n_split}
\alias{n_split}
\title{Split vector in n parts of equal size (or close)}
\usage{
n_split(vec, n)
}
\arguments{
\item{vec}{vector to split}

\item{n}{integer number of parts}
}
\value{
list of size n
}
\description{
If n is greater than the size of the vector, result will have
the size of the vector with one item per element.
}
\examples{
n_split(1:10, 2)
n_split(1:10, 3)

}
