% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{remove_extension}
\alias{remove_extension}
\title{Removes file extension, keeping the path if provided}
\usage{
remove_extension(path)
}
\arguments{
\item{path}{string or vector of strings of file names}
}
\value{
string or vector of strings of file names without extensions
}
\description{
Removes file extension, keeping the path if provided
}
