% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_load_folder}
\alias{s3_load_folder}
\title{List S3 bucket content into data.table}
\usage{
s3_load_folder(
  prefix,
  conf = list(),
  FUN = function(x, files) {     x },
  pattern = "",
  max = Inf,
  quiet = FALSE,
  cl = 1,
  relaunch_times = 1,
  relaunch_cl = 1,
  relaunch_check_fun = is.data.frame,
  ...
)
}
\arguments{
\item{prefix}{string prefix of S3 files to select. Needs to include the URL scheme and
bucket name (s3://bucket_name/)}

\item{conf}{list with 'key', 'secret', 'region', and 'base_url' parameters. If missing, will
look for AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY and AWS_DEFAULT_REGION in environment variables.}

\item{FUN}{function to apply to every file being loaded}

\item{pattern}{string regexp to filter results}

\item{max}{integer max number of records to return, pre-filtering. Default to Inf}

\item{quiet}{logical turn off output messages, default to FALSE}

\item{cl}{integer number of cores to use}
}
\description{
List S3 bucket content into data.table
}
\examples{
\dontrun{
s3_load_folder("s3://hfty-test-bucket/", pattern = "\\\\.rds$")
}

}
