% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_read}
\alias{s3_read}
\title{Read S3 file to R guessing type from extension}
\usage{
s3_read(s3_url, conf = list(), args_read = list(), ...)
}
\arguments{
\item{s3_url}{string S3 URL to read. Needs to include the bucket and protocol (s3://bucket_name/)}

\item{conf}{list with 'key', 'secret', 'region', and 'base_url' parameters. If missing, will
look for AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY and AWS_DEFAULT_REGION in environment variables.}

\item{args_read}{optional, additional arguments passed to read function}

\item{...}{additional arguments passed to aws.s3::save_object}
}
\description{
Read S3 file to R guessing type from extension
}
\examples{
\dontrun{
s3_read("s3://hfty-test-bucket/cars.rds")
s3_read("s3://hfty-test-bucket/cars.csv")
}

}
