% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_save}
\alias{s3_save}
\title{Save R object to S3 guessing type from extension}
\usage{
s3_save(object, s3_url, conf = list(), args_save = list(), quiet = FALSE, ...)
}
\arguments{
\item{object}{current session object to save}

\item{s3_url}{string S3 URL to save to. Needs to include the bucket and protocol (s3://bucket_name/)}

\item{conf}{list with 'key', 'secret', 'region', and 'base_url' parameters. If missing, will
look for AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY and AWS_DEFAULT_REGION in environment variables.}

\item{args_save}{optional, additional arguments passed to save function}

\item{quiet}{logical turn off output messages, default to FALSE}

\item{...}{additional arguments passed to aws.s3::put_object}
}
\description{
Save R object to S3 guessing type from extension
}
\examples{
\dontrun{
s3_save(cars, "s3://hfty-test-bucket/cars.rds")
s3_save(cars, "s3://hfty-test-bucket/cars.csv")
s3_save(cars, "s3://hfty-test-bucket/cars.fwf")
}

}
