% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_save_dt_folder}
\alias{s3_save_dt_folder}
\title{Save data.table to S3 folder split on specific variable(s)}
\usage{
s3_save_dt_folder(
  dt,
  s3_url,
  conf = list(),
  args_save = list(),
  quiet = FALSE,
  ...
)
}
\arguments{
\item{dt}{a data.table}

\item{s3_url}{string S3 URL to save to. Needs to include the bucket and protocol (s3://bucket_name/)
and one or several {glue} strings to be interpolated based on column names}

\item{conf}{list with 'key', 'secret', 'region', and 'base_url' parameters. If missing, will
look for AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY and AWS_DEFAULT_REGION in environment variables.}

\item{args_save}{optional, additional arguments passed to save function}

\item{quiet}{logical turn off output messages, default to FALSE}

\item{...}{additional arguments passed to aws.s3::put_object}
}
\description{
The variables are automatically found from {glue} strings in URL
}
\examples{
\dontrun{
s3_save_dt_folder(dt, "s3://hfty-test-bucket/obj_{date}.rds")
}

}
