% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_dt_folder}
\alias{save_dt_folder}
\title{Save data.table to folder split on specific variable(s)}
\usage{
save_dt_folder(
  dt,
  file,
  threads = parallel::detectCores(),
  compress_level = 9L,
  quiet = FALSE
)
}
\arguments{
\item{dt}{a data.table}

\item{quiet}{logical turn off output messages, default to FALSE}

\item{s3_url}{string S3 URL to save to. Needs to include the bucket and protocol (s3://bucket_name/)
and one or several {glue} strings to be interpolated based on column names}

\item{conf}{list with 'key', 'secret', 'region', and 'base_url' parameters. If missing, will
look for AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY and AWS_DEFAULT_REGION in environment variables.}

\item{args_save}{optional, additional arguments passed to save function}

\item{...}{additional arguments passed to aws.s3::put_object}
}
\description{
The variables are automatically found from {glue} strings in URL
}
\examples{
\dontrun{
save_dt_folder(dt, "dir1/obj_{date}.rds")
}

}
