% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_rds}
\alias{save_rds}
\title{Save to rds using multiple threads (requires xz or pixz)}
\usage{
save_rds(
  x,
  file,
  threads = parallel::detectCores(),
  compress_level = 9L,
  quiet = FALSE
)
}
\arguments{
\item{x}{object to save}

\item{file}{string file path to save to}

\item{threads}{integer number of threads to use, default to parallel::detectCores()}

\item{compress_level}{integer xz compression level from 0 to 9, default to 9 (max)}

\item{quiet}{logical turn off output messages, default to FALSE}
}
\description{
Save to rds using multiple threads (requires xz or pixz)
}
