% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{third_friday_of_month}
\alias{third_friday_of_month}
\title{Find third Friday of month of a given date}
\usage{
third_friday_of_month(date)
}
\arguments{
\item{date}{Date or character representation of date}
}
\value{
Date object of the third Friday of month
}
\description{
Find third Friday of month of a given date
}
