% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{url_scheme}
\alias{url_scheme}
\title{Get URL scheme ('http://', etc.)}
\usage{
url_scheme(path, colonslashes = TRUE)
}
\arguments{
\item{path}{string or vector of strings of file names}

\item{colonslashes}{whether to include '://' and ':', defaults to TRUE}
}
\description{
Get URL scheme ('http://', etc.)
}
